EAIOS-main/backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
websockets==13.1

# Memory Management
mem0ai==0.1.30

# LLM
openai==1.54.0

# MCP
mcp==1.2.1

# Database & Vector Store
chromadb==0.5.20
qdrant-client==1.12.1

# Agent Orchestration
langgraph>=0.2.40
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# Data Processing
pydantic>=2.10.1
pydantic-settings>=2.6.1
python-dotenv==1.0.1

# Utils
httpx==0.27.2
aiofiles==24.1.0
